feat: update progressBar style file, rename button style file
