+++ src/packages/solid/components/ProgressBar/ProgressBar.tsx
@@ -18,6 +18,7 @@
 import type { Component } from 'solid-js';
 import { View, type NodeProps } from '@lightningjs/solid';
 import styles from './ProgressBar.styles.js';
+import type { Tone } from 'types';
 
 export interface ProgressBarProps extends ProgressBarStyleProps, NodeProps {
   /**
@@ -32,6 +33,7 @@ export interface ProgressBarProps extends ProgressBarStyleProps, NodeProps {
    * total height of the component
    */
   height: number;
+  tone: Tone;
 }
 
 export interface ProgressBarStyleProps {
@@ -41,10 +43,17 @@ export interface ProgressBarStyleProps {
 
 const ProgressBar: Component<ProgressBarProps> = (props: ProgressBarProps) => {
   return (
-    <View {...props} style={styles.container}>
+    <View
+      {...props}
+      style={styles.Container}
+      tone={props.tone || styles.tone}
+      {...styles.Container[props.tone || styles.tone]}
+    >
       <View
-        style={styles.progressBar}
+        style={styles.ProgressBar}
         animate
+        tone={props.tone || styles.tone}
+        {...styles.ProgressBar[props.tone || styles.tone]}
         animationSettings={props.animationSettings}
         width={props.width * props.progress}
         borderRadius={props.borderRadius}
